// src/shared/enums/error-type.enum.ts
export enum ERROR_TYPE_ENUM {
  UNAUTHORIZED = "UNAUTHORIZED",
  CONFIGURATION_ERROR = "CONFIGURATION_ERROR",
  NOT_FOUND = "NOT_FOUND",
  BAD_REQUEST = "BAD_REQUEST",
  SERVICE_UNAVAILABLE = "SERVICE_UNAVAILABLE",
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
  DATABASE_ERROR = "DATABASE_ERROR",
  VALIDATION_ERROR = "VALIDATION_ERROR",
  RESOURCE_CONFLICT = "RESOURCE_CONFLICT",
  REQUEST_TIMEOUT = "REQUEST_TIMEOUT",
  FILE_NOT_FOUND = "FILE_NOT_FOUND",
  USER_AGENT_NOT_FOUND = "USER_AGENT_NOT_FOUND",
  FORBIDDEN = "FORBIDDEN",
}